fix(rpc): complete TS type adoption; end duplicated rpc types
Review fix for the "shared types.ts is dead code" finding: the TUI and CLI
local rpc/types.ts are now thin re-exports of @future-os/rpc, making the
shared package the single source of truth for TS RPC payload types and
removing ~425 lines of duplicated definitions.

- shared types.ts gains PermissionLevel (CLI needs it) and legacy
  snake_case aliases on SessionSummary / RecentTerminalAck. The agent
  emits both canonical camelCase and legacy snake_case during the migration
  window, and client code reads a mix, so the shared types model both.
  Fields read as required (updated_at, run_id) are typed required to match
  existing call sites.
- tui/src/rpc/types.ts and cli/src/rpc/types.ts become
  `export type * from "@future-os/rpc"`; all existing `from "./types.js"`
  imports keep working. The unused promptCmd helper is dropped.

Verified: tui+cli tsc clean, 162+108 bun tests pass, both standalone
binaries build.
